[main] Add library UI layer

+++ app/src/main/java/com/example/booknet/apimodels/response/library/LibraryBookDto.kt
@@ -108,5 +108,5 @@ data class LibInfoDto(
     val addDate: Long?,
     @SerializedName("last_chr_count")
     val lastChrCount: Long?,
-    val type: Long?
+    val type: Int?
 )
